% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_normalize.DataFrame.R
\name{custom_normalize.DataFrame}
\alias{custom_normalize.DataFrame}
\title{Custom Data Frame Normalization Function}
\usage{
custom_normalize.DataFrame(df, method = "standardize")
}
\arguments{
\item{df}{A data frame to be normalized.}

\item{method}{The normalization method to be applied. Default is "standardize".
Possible values: "standardize", "min_max", "log_transform", "robust", "z_score", "box_cox".}
}
\value{
Returns a data frame with numeric columns normalized or standardized according to the specified method.
}
\description{
This function performs normalization or standardization on numeric columns of a data frame.
}
\details{
This function separates numeric and non-numeric columns. It then applies the chosen normalization method only to the numeric columns.
The supported normalization methods are:
  - "standardize": Standardization (centering and scaling) of numeric columns.
  - "min_max": Min-Max Normalization of numeric columns.
  - "log_transform": Log transformation of numeric columns.
  - "robust": Robust standardization of numeric columns.
  - "z_score": Z-Score normalization of numeric columns.
}
\examples{
# Example usage:
# Creating a sample data frame
my_data_frame <- data.frame(
  NumericColumn1 = c(10, 20, 30, 40, 50),
  NumericColumn2 = c(1.5, 2.5, 3.5, 4.5, 5.5),
  CategoricalColumn = c("A", "B", "A", "B", "A")
)

# Changing class of data frame to DataFrame
my_data_frame <- DataFrame(my_data_frame)

# Standardize numeric columns in a data frame
result_standardize <- custom_normalize.DataFrame(my_data_frame, method = "standardize")


# Min-Max normalize numeric columns in a data frame
result_min_max <- custom_normalize.DataFrame(my_data_frame, method = "min_max")

}
