% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePlot.DataFrame.R
\name{GeneratePlot.DataFrame}
\alias{GeneratePlot.DataFrame}
\title{GeneratePlot function}
\usage{
GeneratePlot.DataFrame(data, column, plot_type = "distribution")
}
\arguments{
\item{data}{A data frame.}

\item{column}{The column for which the plot should be generated.}

\item{plot_type}{The type of plot to generate. Options include:
\code{"distribution"}: Generate a distribution plot.
\code{"boxplot"}: Generate a boxplot.}
}
\value{
A plot object (either distribution plot or boxplot).
}
\description{
This function generates either a distribution plot or a boxplot for a specified column in a data frame.
}
\examples{
\dontrun{
# Generate distribution plot for the "value" column
plot_dist <- GeneratePlot(data, column = "value", plot_type = "distribution")
plot_dist

# Generate boxplot for the "value" column
plot_box <- GeneratePlot(data, column = "value", plot_type = "boxplot")
plot_box
}

}
