\name{as.data.frame.chron}
\alias{as.data.frame.chron}
\alias{as.data.frame.dates}
\alias{as.data.frame.times}
\alias{as.data.frame.Date}
\alias{as.data.frame.POSIXct}
\alias{as.data.frame.numeric_version}
\docType{package}
\title{
  Coerce to a Data Frame
}
\description{
  These functions are copies of \code{\link{as.data.frame.vector}}.
}
\seealso{
  \code{\link{as.data.frame.vector}},
  \code{\link[chron]{as.data.frame.chron}},
  \code{\link[chron]{as.data.frame.dates}},
  \code{\link[chron]{as.data.frame.times}},
  \code{\link[base]{as.data.frame.Date}},
  \code{\link[base]{as.data.frame.POSIXct}},
  \code{\link[base]{as.data.frame.numeric_version}}.
}
\keyword{classes}
\keyword{methods}
